Move hooks to `plugin/cli`
