+++ src/core/notices/index.ts
@@ -1,3 +1,9 @@
+/**
+ * 通知相关操作
+ *
+ * @module
+ */
+
 export { diff, merge } from './comparer.js'
 export { fetch_all_sources } from './fetcher.js'
 export { read_json, write_json } from './saver.js'
