+++ src/plugin/cli/examples/cli.ts
@@ -1,0 +1,8 @@
+import { hook, update_notices } from '../../../core/index.js'
+import { add_hook } from '../index.js'
+
+add_hook.verbose(hook)
+add_hook.progress_bar(hook)
+add_hook.recent_filter(hook, 90)
+
+await update_notices()
